+++ baseapp/iostats_nop.go
@@ -1,0 +1,8 @@
+// +build !linux
+
+package baseapp
+
+import "github.com/line/ostracon/libs/log"
+
+func logIoStats(logger log.Logger) {
+}