+++ x/auth/types/params.go
@@ -15,7 +15,7 @@ const (
 	DefaultTxSizeCostPerByte      uint64 = 10
 	DefaultSigVerifyCostED25519   uint64 = 590
 	DefaultSigVerifyCostSecp256k1 uint64 = 1000
-	DefaultValidSigBlockPeriod    uint64 = 100
+	DefaultValidSigBlockPeriod    uint64 = 3600 // 1 hour for 1 second block interval
 )
 
 // Parameter keys