+++ x/consortium/genesis.go
@@ -1,0 +1,43 @@
+package consortium
+
+import (
+	sdk "github.com/line/lbm-sdk/types"
+	"github.com/line/lbm-sdk/x/consortium/keeper"
+	"github.com/line/lbm-sdk/x/consortium/types"
+
+	stakingtypes "github.com/line/lbm-sdk/x/staking/types"
+)
+
+func InitGenesis(ctx sdk.Context, keeper keeper.Keeper, sk types.StakingKeeper, data *types.GenesisState) error {
+	keeper.SetParams(ctx, data.Params)
+
+	validatorAuths := data.ValidatorAuths
+	if keeper.GetEnabled(ctx) && len(validatorAuths) == 0 {
+		// Allowed validators must exist if the module is enabled,
+		// so it should be the very first block of the chain.
+		// We gather the information from staking module.
+		sk.IterateValidators(ctx, func(_ int64, addr stakingtypes.ValidatorI) (stop bool) {
+			auth := &types.ValidatorAuth{
+				OperatorAddress: addr.GetOperator().String(),
+				CreationAllowed: true,
+			}
+			validatorAuths = append(validatorAuths, auth)
+			return false
+		})
+	}
+
+	for _, auth := range validatorAuths {
+		if err := keeper.SetValidatorAuth(ctx, auth); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func ExportGenesis(ctx sdk.Context, keeper keeper.Keeper) *types.GenesisState {
+	return &types.GenesisState{
+		Params:         keeper.GetParams(ctx),
+		ValidatorAuths: keeper.GetValidatorAuths(ctx),
+	}
+}