+++ x/feegrant/simulation/genesis_test.go
@@ -1,0 +1,40 @@
+package simulation_test
+
+import (
+	"encoding/json"
+	"math/rand"
+	"testing"
+
+	"github.com/line/lbm-sdk/x/feegrant/types"
+	"github.com/stretchr/testify/require"
+
+	"github.com/line/lbm-sdk/simapp"
+	"github.com/line/lbm-sdk/types/module"
+	simtypes "github.com/line/lbm-sdk/types/simulation"
+	"github.com/line/lbm-sdk/x/feegrant/simulation"
+)
+
+func TestRandomizedGenState(t *testing.T) {
+	app := simapp.Setup(false)
+
+	s := rand.NewSource(1)
+	r := rand.New(s)
+
+	accounts := simtypes.RandomAccounts(r, 3)
+
+	simState := module.SimulationState{
+		AppParams:    make(simtypes.AppParams),
+		Cdc:          app.AppCodec(),
+		Rand:         r,
+		NumBonded:    3,
+		Accounts:     accounts,
+		InitialStake: 1000,
+		GenState:     make(map[string]json.RawMessage),
+	}
+
+	simulation.RandomizedGenState(&simState)
+	var feegrantGenesis types.GenesisState
+	simState.Cdc.MustUnmarshalJSON(simState.GenState[types.ModuleName], &feegrantGenesis)
+
+	require.Len(t, feegrantGenesis.Allowances, len(accounts)-1)
+}